// google-cloud-ai-in-action/app/types/pagination.tsx
import  { Dispatch, SetStateAction, useEffect, useState } from "react";

export type PaginationType = {
    entryCount: number;
    entryCountPerPage: number;
    currentPageIndex: number;
}

export type PaginationWrapperType = {
    pagination: PaginationType;
    setPagination: Dispatch<SetStateAction<PaginationType | null>>;
}